@@ -24,8 +24,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <source>11</source>
-                    <target>11</target>
+                    <source>16</source>
+                    <target>16</target>
                 </configuration>
             </plugin>
             <plugin>
@@ -44,6 +44,10 @@
                                 <relocation>
                                     <pattern>com.jeff_media.morepersistentdatatypes</pattern>
                                     <shadedPattern>me.will0mane.plugins.adventure.morepersistentdatatypes</shadedPattern>
+                                </relocation>
+                                <relocation>
+                                    <pattern>com.jeff_media.customblockdata</pattern>
+                                    <shadedPattern>me.will0mane.plugins.adventure.customblockdata</shadedPattern>
                                 </relocation>
                             </relocations>
                         </configuration>
@@ -72,6 +76,10 @@
             <id>jeff-media-public</id>
             <url>https://hub.jeff-media.com/nexus/repository/jeff-media-public/</url>
         </repository>
+        <repository>
+            <id>jitpack.io</id>
+            <url>https://jitpack.io</url>
+        </repository>
     </repositories>
 
     <dependencies>
@@ -86,6 +94,29 @@
             <artifactId>MorePersistentDataTypes</artifactId>
             <version>2.3.1</version>
             <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>RELEASE</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.jetbrains</groupId>
+            <artifactId>annotations</artifactId>
+            <version>RELEASE</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.jeff_media</groupId>
+            <artifactId>CustomBlockData</artifactId>
+            <version>2.2.0</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.github.Will0mane</groupId>
+            <artifactId>MongoWillo</artifactId>
+            <version>2.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
 </project>
