@@ -1,13 +1,7 @@
 package me.will0mane.plugins.adventure.systems.items.abilities;
 
-import me.will0mane.plugins.adventure.systems.items.abilities.data.InteractAbility;
-import net.minecraft.sounds.SoundEffects;
-import org.bukkit.Bukkit;
-
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
-import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 
@@ -37,6 +31,7 @@ public abstract class ItemAbility<T extends AbilityData> {
 
     public abstract String activationMethodName();
 
+    @SuppressWarnings("unchecked")
     public void triggerWithArgs(Class<?> data, Object[] arguments){
         try {
             trigger((T) data.getDeclaredConstructors()[0].newInstance(arguments));
